Initial work on MSBuild SignTool Task Wrapper
