+++ src/Microsoft.DotNet.SignTool/src/BatchSignInput.cs
