+++ src/Microsoft.DotNet.SignTool/src/BatchSignUtil.cs
