+++ src/Microsoft.DotNet.SignTool/src/ContentMap.cs
