+++ src/Microsoft.DotNet.SignTool/src/ContentUtil.cs
