+++ src/Microsoft.DotNet.SignTool/src/FileName.cs
