+++ src/Microsoft.DotNet.SignTool/src/FileSignInfo.cs
