+++ src/Microsoft.DotNet.SignTool/src/IHost.cs
