+++ src/Microsoft.DotNet.SignTool/src/JsonTypes.cs
