+++ src/Microsoft.DotNet.SignTool/src/OrchestratedBatchSignInput.cs
