+++ src/Microsoft.DotNet.SignTool/src/PathUtil.cs
