+++ src/Microsoft.DotNet.SignTool/src/SignInfo.cs
