+++ src/Microsoft.DotNet.SignTool/src/SignTool.RealSignTool.cs
