+++ src/Microsoft.DotNet.SignTool/src/SignTool.SignToolBase.cs
