+++ src/Microsoft.DotNet.SignTool/src/SignTool.TestSignTool.cs
