+++ src/Microsoft.DotNet.SignTool/src/SignTool.cs
