+++ src/Microsoft.DotNet.SignTool/src/ZipData.cs
