+++ src/Microsoft.DotNet.SignTool/src/ZipPart.cs
